xntop-importer/requirements.txt
psycopg2-binary
requests
python-dotenv
tenacity
